requirements.txt
keyboard==0.13.5
notify-py==0.3.42
Pillow==9.4.0
PyAutoGUI==0.9.53
pyperclip==1.8.2
PySimpleGUI==4.60.4
pystray==0.19.4
requests==2.28.2